ml-service/requirements.txt
flask
flask-cors
psycopg2-binary
python-dotenv
pandas
numpy
scikit-learn